transform wip

+++ Lumos/src/Scripting/Bindings/CameraBindings/Camera2DBinding.cpp
@@ -58,7 +58,7 @@ namespace Lumos
 				}
 			}
 		}
-		camera = new Camera2D(fov, n,1.0f);
+		camera = new Camera2D((uint)fov, (uint)n,1.0f);
 	}
 
 	Camera2DBinding::~Camera2DBinding()
